Added transition names
The simulation now also returns a list of names of the consecutive transitions.  Prerequisite is the rate_name_dict as a second return value of initialize.rate_assignment.

The gillespie algorithm code was simplified which speeds it up (at the moment it is faster than cython implementation).

The cython implementation lacks the constrcution of the transtion name list.

Finished with all tests passed.