mistake in TCSPC simulation
Earlier, discarding of the TCSPC simulation was done due to no agreement of measured triplet lifetime with expected triplet lifetime. After constructing a new simulation based on a waiting array it became clear that it should be working in the original simulation and something else is going wrong. Indeed, it was the wrong ahndling of excitations (by pulse).
This upload is to save the status of the complicated but probably unnecessary new simulation approach that is from now getting back-replaced by the original one.